Compile  desktopMain om te testen, FTPservice werkt op pc maar nog niet op android omdat ik nog coroutines moet toevoegen. Als je wilt veranderen welke file wordt gedownload moet je de index van de downloadfile veranderen in commonMain\kotlin\ftptest.kt, op line 40
feat: Add ftptest composable to App

This commit adds a new composable function called `ftptest` to the `App` module. The `ftptest` composable receives an instance of `FtpClient` and establishes an FTP connection with the server. It then retrieves a list of files from a specific directory and downloads a file from the server if the directory is not empty. The downloaded file is stored in the `.downloads` directory.